Move StringReplaceWithLowercase() from tools to translations
