+++ src/engine/tools.h
@@ -54,8 +54,6 @@ std::string StringUpper( std::string str );
 
 std::vector<std::string> StringSplit( const std::string & str, const char sep );
 
-// Function to replace the pattern in workString with patternReplacement. Here the patternReplacement is converted to lowercase except for the first word in a sentence.
-void StringReplaceWithLowercase( std::string & workString, const char * pattern, const std::string & patternReplacement );
 void StringReplace( std::string & dst, const char * pred, const std::string & src );
 
 template <typename T, typename = typename std::enable_if<std::is_integral<T>::value>::type>
